backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
sqlalchemy==2.0.28
python-dotenv==1.0.1
ftrack-python-api==2.5.0
pydantic==2.6.4
pandas==2.2.1
alembic==1.13.1
